Move logs to dedicated directory
