+++ packages/core/src/index.tsx
@@ -15,7 +15,8 @@ import { InternalLog } from './InternalLog';
 import { ProxyConfiguration, ProxyType } from './ProxyConfiguration';
 import { SdkVerbosity } from './SdkVerbosity';
 import { TrackingConsent } from './TrackingConsent';
-import { DdLogs, DdTrace, DdRum } from './foundation';
+import { DdTrace, DdRum } from './foundation';
+import { DdLogs } from './logs/DdLogs';
 import { RumActionType, ErrorSource } from './types';
 
 export {
