+++ packages/core/src/types.tsx
@@ -60,39 +60,6 @@ export type DdSdkType = {
 };
 
 /**
- * The entry point to use Datadog's Logs feature.
- */
-export type DdLogsType = {
-    /**
-     * Send a log with level debug.
-     * @param message: The message to send.
-     * @param context: The additional context to send.
-     */
-    debug(message: string, context?: object): Promise<void>;
-
-    /**
-     * Send a log with level info.
-     * @param message: The message to send.
-     * @param context: The additional context to send.
-     */
-    info(message: string, context?: object): Promise<void>;
-
-    /**
-     * Send a log with level warn.
-     * @param message: The message to send.
-     * @param context: The additional context to send.
-     */
-    warn(message: string, context?: object): Promise<void>;
-
-    /**
-     * Send a log with level error.
-     * @param message: The message to send.
-     * @param context: The additional context to send.
-     */
-    error(message: string, context?: object): Promise<void>;
-};
-
-/**
  * The entry point to use Datadog's Trace feature.
  */
 export type DdTraceType = {
